+++ trunk/BaseletElements/src/com/umlet/element/experimental/settings/SettingsClass.java
@@ -14,7 +14,7 @@ import com.umlet.element.experimental.settings.facets.TemplateClass;
 public class SettingsClass extends Settings {
 
 	@Override
-	public XValues getXValues(float y, int height, int width) {
+	public XValues getXValues(double y, int height, int width) {
 		return new XValues(0, width);
 	}
