+++ trunk/BaseletElements/src/com/umlet/element/experimental/settings/SettingsInterface.java
@@ -13,7 +13,7 @@ import com.umlet.element.experimental.settings.facets.DefaultGlobalFacet.Element
 public class SettingsInterface extends Settings {
 
 	@Override
-	public XValues getXValues(float y, int height, int width) {
+	public XValues getXValues(double y, int height, int width) {
 		return new XValues(0, width);
 	}
