+++ trunk/BaseletElements/src/com/umlet/element/experimental/settings/facets/SeparatorLine.java
@@ -26,7 +26,7 @@ public class SeparatorLine implements Facet {
 		if (setHAlignToLeftAfterLine) {
 			propConfig.sethAlign(AlignHorizontal.LEFT);
 		}
-		float linePos = propConfig.getDividerPos(drawer.textHeight());
+		double linePos = propConfig.getDividerPos(drawer.textHeight());
 		XValues xPos = propConfig.getXLimits(linePos);
 		drawer.drawLine(xPos.getLeft()+1, linePos, xPos.getRight()-1, linePos);
 		propConfig.addToYPos(H_SPACE);
